# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: AgentGate Control Plane API
  version: 0.3.0
  description: |
    API-first control plane for policy-governed agent operations over business APIs.
    This spec freezes contract shape for partner-facing pilot endpoints.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /gateway/access/tokens:
    post:
      tags: [Auth]
      summary: Create or update an access token (admin only)
      parameters:
        - $ref: '#/components/parameters/XGatewayToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessTokenCreate'
            examples:
              tenantOperator:
                summary: Create a tenant-operator token
                value:
                  token: tenant-a-token
                  subject_id: tenant-a-op
                  tenant_id: tenant-a
                  role: tenant-operator
                  scopes: [gateway:read, gateway:write, gateway:plan]
                  enabled: true
              admin:
                summary: Create an admin token with expiry
                value:
                  token: ops-admin-token
                  subject_id: ops-admin
                  role: admin
                  scopes: [gateway:*]
                  enabled: true
                  expires_at: "2026-12-31T23:59:59+00:00"
      responses:
        '200':
          description: Token created/updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /gateway/servers:
    get:
      tags: [Gateway]
      summary: List servers
      parameters:
        - $ref: '#/components/parameters/XGatewayToken'
      responses:
        '200':
          description: Server list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RegisteredServer'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Gateway]
      summary: Register server
      parameters:
        - $ref: '#/components/parameters/XGatewayToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterServerRequest'
            examples:
              sseServer:
                summary: Register an SSE transport server
                value:
                  server_id: srv-1
                  name: github-mcp
                  tenant_id: tenant-a
                  transport: sse
                  endpoint: https://example.com/sse
                  tags: [finance]
                  weight: 5
                  priority: 20
              stdioServer:
                summary: Register a stdio transport server
                value:
                  server_id: srv-local
                  name: local-mcp
                  tenant_id: tenant-a
                  transport: stdio
                  endpoint: ""
                  tags: [dev]
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredServer'
              example:
                server_id: srv-1
                name: github-mcp
                tenant_id: tenant-a
                transport: sse
                endpoint: https://example.com/sse
                tags: [finance]
                healthy: true
                weight: 5
                priority: 20
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /gateway/policy/rules:
    post:
      tags: [Policy]
      summary: Create policy rule
      parameters:
        - $ref: '#/components/parameters/XGatewayToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyRuleCreate'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /gateway/routes/dry-run:
    post:
      tags: [Gateway]
      summary: Dry-run route selection + policy decision
      parameters:
        - $ref: '#/components/parameters/XGatewayToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteDryRunRequest'
            examples:
              allowPath:
                summary: Dry-run that matches an allow rule
                value:
                  tenant_id: tenant-a
                  action: plan
                  server_id: srv-1
                  strategy: preferred
              denyPath:
                summary: Dry-run with unknown server (denied)
                value:
                  tenant_id: tenant-a
                  action: plan
                  server_id: srv-nonexistent
                  strategy: preferred
              weightedPlan:
                summary: Weighted strategy selection
                value:
                  tenant_id: tenant-a
                  action: plan
                  strategy: weighted
                  preferred_server_id: srv-1
      responses:
        '200':
          description: Decision result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDryRunResponse'
              examples:
                allowed:
                  summary: Policy allows the request
                  value:
                    event_id: "550e8400-e29b-41d4-a716-446655440000"
                    decided_at: "2026-03-11T12:00:00+00:00"
                    actor:
                      subject_id: tenant-a-op
                      role: tenant-operator
                      tenant_id: tenant-a
                    request:
                      tenant_id: tenant-a
                      action: plan
                    allowed: true
                    decision: allow
                    reason: "allow: matched rule allow-plan"
                    matched_rule: allow-plan
                    selected_server_id: srv-1
                    server_healthy: true
                    strategy: preferred
                denied:
                  summary: Policy denies — server not registered
                  value:
                    event_id: "550e8400-e29b-41d4-a716-446655440001"
                    decided_at: "2026-03-11T12:00:01+00:00"
                    actor:
                      subject_id: tenant-a-op
                      role: tenant-operator
                      tenant_id: tenant-a
                    request:
                      tenant_id: tenant-a
                      action: plan
                    allowed: false
                    decision: deny
                    reason: "deny: target server not registered"
                    matched_rule: null
                    selected_server_id: null
                    server_healthy: null
                    strategy: preferred
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /gateway/audit:
    get:
      tags: [Gateway]
      summary: Export audit log of dry-run decisions
      parameters:
        - $ref: '#/components/parameters/XGatewayToken'
        - in: query
          name: tenant_id
          required: false
          schema:
            type: string
          description: Filter by tenant
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: Audit entries (most recent first)
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RouteDryRunResponse'
              example:
                items:
                  - event_id: "550e8400-e29b-41d4-a716-446655440000"
                    decided_at: "2026-03-11T12:00:00+00:00"
                    actor:
                      subject_id: tenant-a-op
                      role: tenant-operator
                      tenant_id: tenant-a
                    request:
                      tenant_id: tenant-a
                      action: plan
                    allowed: true
                    decision: allow
                    reason: "allow: matched rule allow-plan"
                    matched_rule: allow-plan
                    selected_server_id: srv-1
                    server_healthy: true
                    strategy: preferred
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  parameters:
    XGatewayToken:
      in: header
      name: x-gateway-token
      required: true
      schema:
        type: string
      description: Gateway access token
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status: { type: string, example: ok }
    AccessTokenCreate:
      type: object
      required: [token, subject_id, role]
      properties:
        token: { type: string }
        subject_id: { type: string }
        tenant_id: { type: string, nullable: true }
        role:
          type: string
          enum: [admin, tenant-operator, reader]
        scopes:
          type: array
          items: { type: string }
        enabled:
          type: boolean
          default: true
        expires_at:
          type: string
          nullable: true
          description: ISO-8601 UTC timestamp after which token is rejected
    RegisterServerRequest:
      type: object
      required: [server_id, name, tenant_id, transport, endpoint]
      properties:
        server_id: { type: string }
        name: { type: string }
        tenant_id: { type: string }
        transport:
          type: string
          enum: [stdio, sse, streamable-http, ws]
        endpoint: { type: string }
        tags:
          type: array
          items: { type: string }
        weight:
          type: integer
          minimum: 1
          maximum: 100
        priority:
          type: integer
          minimum: 1
          maximum: 1000
    RegisteredServer:
      allOf:
        - $ref: '#/components/schemas/RegisterServerRequest'
        - type: object
          properties:
            healthy: { type: boolean }
    PolicyRuleCreate:
      type: object
      required: [name, effect, actions, resources]
      properties:
        name: { type: string }
        effect:
          type: string
          enum: [allow, deny]
        actions:
          type: array
          items: { type: string }
        resources:
          type: array
          items: { type: string }
        tenants:
          type: array
          items: { type: string }
          nullable: true
    RouteDryRunRequest:
      type: object
      required: [tenant_id, action]
      properties:
        tenant_id: { type: string }
        action: { type: string }
        server_id: { type: string, nullable: true }
        preferred_server_id: { type: string, nullable: true }
        strategy:
          type: string
          enum: [preferred, weighted, failover]
          default: preferred
    RouteDryRunResponse:
      type: object
      properties:
        event_id: { type: string, description: Unique decision event id }
        decided_at: { type: string, description: ISO-8601 UTC timestamp for decision }
        actor:
          type: object
          properties:
            subject_id: { type: string }
            role: { type: string }
            tenant_id: { type: string, nullable: true }
        request:
          type: object
          properties:
            tenant_id: { type: string }
            action: { type: string }
        allowed: { type: boolean }
        decision:
          type: string
          enum: [allow, deny]
        reason: { type: string }
        matched_rule: { type: string, nullable: true }
        selected_server_id: { type: string, nullable: true }
        server_healthy: { type: boolean, nullable: true }
        strategy:
          type: string
          enum: [preferred, weighted, failover]
    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
  responses:
    Unauthorized:
      description: Missing/invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Scope/role/tenant violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Payload failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'